Change macro for creating plugins.
Qt MOC does not detect Q_OBJECT inside custom MACRO. Consequently, if
DTK_DECLARE_PLUGIN macro was used in a header file that does not contain
Q_OBJECT macro somewhere, the resulting class from the macro was not
moced.

Now, a pure virtual class is declared in dtkCore and the macro is just
an inheritance of this class that has been moced.